move AmmoniteClassFinder to connect.client package
