+++ connector/connect/client/jvm/src/test/scala/org/apache/spark/sql/connect/client/CheckConnectJvmClientCompatibility.scala
@@ -335,13 +335,13 @@ object CheckConnectJvmClientCompatibility {
       ),
       // ConnectRepl
       ProblemFilters.exclude[MissingClassProblem](
-        "org.apache.spark.sql.application.AmmoniteClassFinder" // developer API
-      ),
-      ProblemFilters.exclude[MissingClassProblem](
         "org.apache.spark.sql.application.ConnectRepl" // developer API
       ),
       ProblemFilters.exclude[MissingClassProblem](
         "org.apache.spark.sql.application.ConnectRepl$" // developer API
+      ),
+      ProblemFilters.exclude[MissingClassProblem](
+        "org.apache.spark.sql.connect.client.AmmoniteClassFinder" // developer API
       ),
 
       // SparkSession
